Migrate the project to Autotools
This standardizes the project's build and install process by adapting
the well-known Autotools system. This simply migrates the previous
build.sh script into Autotools form. This does not attempt to clean up
or restructure anything to become more friendly to Autotools.

In order to better the integration of Autotools, the directories and
files should become more "friendly." The way things are currently laid
out, it requires removing and copying all over the place. What should
probably be done is to store only what differs between similar items,
and just move or replace strings, for instance, the index.themes.